+++ Changelog.md
@@ -1,6 +1,15 @@
 ## Release History
 
 
+### 7.4.2 ( 2024-01-30)
+- [x] Dependencies updates. @peter.naydenov/walk to version 4.2.3;
+- [x] Dev dependencies updates. Chai to version 5.0.3;
+- [x] Dev dependencies updates. C8 to version 9.1.0;
+- [x] Folder 'dist' was added to the project. Includes commonjs, umd and esm versions of the library;
+- [x] Package.json: "exports" section was added. Allows you to use package as commonjs or es6 module without additional configuration;
+- [x] Rollup was added to the project. Used to build the library versions;
+
+
 
 ### 7.4.1 ( 2024-01-10)
 - [x] Dependencies updates. @peter.naydenov/walk to version 4.2.1;